Restructure directories to match new architecture
`src/actors` no longer contained multiple actors. It contained `Actor`
and then a library of roles that actor could take on. Those roles feel
like a natural directory. It then makes a bit more sense to me if we
move `Actor` to the top level.

This does mean that `src/roles/*.rs` is implementation details of
`src/actor.rs`, which is a little weird in terms of the module
structure, but as `roles` is a private module this isn't misleading.